// v2-src/types.ts
export const Operators = [
  // Canonical operator names
  'AND',
  'OR',
  'EQUAL',
  'NOT_EQUAL',
  'PLUS',
  'SUBTRACT',
  'MULTIPLY',
  'DIVIDE',
  'GREATER_THAN',
  'LESS_THAN',
  'CONDITIONAL',
  'REGEX',
  'OBJECT_PROPERTIES',
  'STRING_SUBSTITUTION',
  'SPLIT',
  'COUNT',
  'GET',
  'POST',
  'SQL',
  'GRAPHQL',
  'BUILD_OBJECT',
  'MATCH',
  'CUSTOM_FUNCTIONS',
  'PASSTHRU',
] as const

export type Operator = (typeof Operators)[number]
export type OperatorAlias = keyof typeof operatorAliases
export type OperatorAliases = Record<string, Operator>

export type Fragments = Record<string, Fragment>
// ...
export type OutputType = 'string' | 'number' | 'boolean' | 'bool' | 'array'

export interface OperatorNode {
  operator: OperatorAlias
  outputType?: OutputType
  children?: EvaluatorNode[]
  fallback?: EvaluatorNode
  useCache?: boolean
  // Additional parameters and Alias nodes
  [key: string]: EvaluatorNode
}

export interface FragmentNode {
  fragment: string
  parameters?: Record<string, EvaluatorNode>
  // For parameters at the root level
  [key: string]: EvaluatorNode
}

export interface FragmentData {
  description?: string
  parameters?: FragmentParameterMetadata[]
  textColor?: string
  backgroundColor?: string
}

export interface FunctionDefinition {
  function: UnknownFunction
  description?: string
  argsDefault?: unknown[]
  inputDefault?: Record<string, unknown>
  textColor?: string
  backgroundColor?: string
}

export type Fragment =
  | (EvaluatorNode & {
      metadata?: FragmentData
    })
  | null

export type EvaluatorOutput =
  | string
  | boolean
  | number
  | object
  | null
  | undefined
  | Array<EvaluatorOutput>

export type EvaluatorNode = OperatorNode | FragmentNode | EvaluatorOutput